Final animations+without coping
