+++ Abyss/Assets/1 - Art/Animations/Spelunker_anim (1).fbx
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:1e03cb731d4457d504d2c0f1c1f3f7cfce8daa7a1c1f3f0f0f6cd31d9b38411a
+size 4261820
