+++ Abyss/Assets/3 - Prefabs/Player/Player.prefab
@@ -309,7 +309,7 @@ GameObject:
   - component: {fileID: 3337618906554173024}
   - component: {fileID: 1258608492262260742}
   - component: {fileID: 77520714226835381}
-  - component: {fileID: 2852131865764617005}
+  - component: {fileID: 8112239056880653029}
   m_Layer: 0
   m_Name: Player
   m_TagString: Untagged
@@ -647,7 +647,7 @@ MonoBehaviour:
   - {fileID: 2361723306473841219}
   - {fileID: 5889561739995210761}
   - {fileID: 408416460547995745}
---- !u!114 &2852131865764617005
+--- !u!114 &8112239056880653029
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -656,7 +656,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 2457687564484327083}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: cc198841733c35d43a148bad86ff75ee, type: 3}
+  m_Script: {fileID: 11500000, guid: d897a33717f0e4e4388b6fa88fd71418, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   climbing: {fileID: 6776748995389974420}
@@ -1359,6 +1359,9 @@ PrefabInstance:
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
     m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: -6395168040719377692, guid: 66df67491ad737248a6b36d97c39e299, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 4904735155650731691}
     - targetCorrespondingSourceObject: {fileID: 4949379128218185088, guid: 66df67491ad737248a6b36d97c39e299, type: 3}
       insertIndex: -1
       addedObject: {fileID: 8179672984863828083}
@@ -1693,11 +1696,38 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
---- !u!4 &7737116220751323180 stripped
-Transform:
-  m_CorrespondingSourceObject: {fileID: 2801032980297941759, guid: 66df67491ad737248a6b36d97c39e299, type: 3}
+--- !u!1 &7691948184719210551 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: -6395168040719377692, guid: 66df67491ad737248a6b36d97c39e299, type: 3}
   m_PrefabInstance: {fileID: 5584611499581350611}
   m_PrefabAsset: {fileID: 0}
+--- !u!54 &4904735155650731691
+Rigidbody:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7691948184719210551}
+  serializedVersion: 4
+  m_Mass: 1
+  m_Drag: 0
+  m_AngularDrag: 0.05
+  m_CenterOfMass: {x: 0, y: 0, z: 0}
+  m_InertiaTensor: {x: 1, y: 1, z: 1}
+  m_InertiaRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ImplicitCom: 1
+  m_ImplicitTensor: 1
+  m_UseGravity: 1
+  m_IsKinematic: 1
+  m_Interpolate: 0
+  m_Constraints: 0
+  m_CollisionDetection: 0
 --- !u!4 &8913452633810940536 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 3905315097887215787, guid: 66df67491ad737248a6b36d97c39e299, type: 3}
@@ -3043,9 +3073,6 @@ PrefabInstance:
     - targetCorrespondingSourceObject: {fileID: 4949379128218185088, guid: 66df67491ad737248a6b36d97c39e299, type: 3}
       insertIndex: -1
       addedObject: {fileID: 427685524750911686}
-    - targetCorrespondingSourceObject: {fileID: 4949379128218185088, guid: 66df67491ad737248a6b36d97c39e299, type: 3}
-      insertIndex: -1
-      addedObject: {fileID: 3678298642744047733}
     - targetCorrespondingSourceObject: {fileID: 6509733414440294620, guid: 66df67491ad737248a6b36d97c39e299, type: 3}
       insertIndex: -1
       addedObject: {fileID: 8364233638092165265}
@@ -3676,27 +3703,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   animatedRb: {fileID: 8179672984863828083}
-  isSyncing: 1
---- !u!114 &3678298642744047733
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2833526332457665443}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0611482b4467b2a4590f5b2ad7c9f41a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  sourceRoot: {fileID: 7737116220751323180}
-  targetRoot: {fileID: 4981627221369662684}
-  createMissingTargets: 0
-  copyPosition: 1
-  copyRotation: 1
-  smoothTime: 0
-  skipIfTargetHasRigidbody: 1
-  axisMask: {x: 1, y: 1, z: 1}
+  isSyncing: 0
 --- !u!1 &4155194333580228351 stripped
 GameObject:
   m_CorrespondingSourceObject: {fileID: 6509733414440294620, guid: 66df67491ad737248a6b36d97c39e299, type: 3}
@@ -4037,11 +4044,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   animatedRb: {fileID: 390649620854885139}
   isSyncing: 1
---- !u!4 &4981627221369662684 stripped
-Transform:
-  m_CorrespondingSourceObject: {fileID: 2801032980297941759, guid: 66df67491ad737248a6b36d97c39e299, type: 3}
-  m_PrefabInstance: {fileID: 7204931711383612963}
-  m_PrefabAsset: {fileID: 0}
 --- !u!1 &5565172355861236099 stripped
 GameObject:
   m_CorrespondingSourceObject: {fileID: -5852856583299962976, guid: 66df67491ad737248a6b36d97c39e299, type: 3}
@@ -4220,7 +4222,7 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 5565172355861236099}
-  m_Enabled: 1
+  m_Enabled: 0
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 0611482b4467b2a4590f5b2ad7c9f41a, type: 3}
   m_Name: 
@@ -4411,7 +4413,7 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 5791584840296899866}
-  m_Enabled: 1
+  m_Enabled: 0
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 0611482b4467b2a4590f5b2ad7c9f41a, type: 3}
   m_Name:
